+++ apps/api/src/app/user/update-user-setting.dto.ts
@@ -57,6 +57,10 @@ export class UpdateUserSettingDto {
 
   @IsArray()
   @IsOptional()
+  'filters.assetClasses'?: string[];
+
+  @IsArray()
+  @IsOptional()
   'filters.tags'?: string[];
 
   @IsBoolean()